Experimental patch (committed to objc2 branch only for now, though it's not at all specific to objective c 2.0 - I just wasn't sure if it was too disruptive to push to the trunk) extending previous work on super calls to cover class methods.
This involved adding a fair bit to the output of declareClass - if anyone sees a better way to do this, I'd love to hear about it.  One thing I thought about was using an instance of the form:

instance ClassAndObject meta $(metaClsType) => ClassObject meta
	where classObject = unsafeGetMetaclassForClass classObject

	I believe this would have "worked" with slightly less clutter in the namespace, but it would require adding FlexibleContexts to the LANGUAGE pragma
of every source file that calls declareClass, which seemed even more disruptive.