Placed 'c-sources' line back into cabal file.  I thought this had fixed ghc-6.10 builds, but it turns out the problem was deeper than I thought.
Essentially, this line was fooling Cabal into including dist/build/HOC_cbits.o in the built library.  Under earlier versions of GHC, the ghc call that the cabal generates to handle this "C source" is silently ignored because it is calling 'ghc -c' on an object file.  With ghc 6.10.1, the call dies violently with "no input files".

I have dug around a bit in the cabal api, but not yet found any way of achieving the desired effect, so for now I'm going to have to re-open issue 10 and put the line back.