Have another shot at fixing the cabal build (issue #12).
HOC_cbits.o is passed in via c-sources; Setup.hs now contains a hack that prevents the
"ghc: no input files" problem.

To be exact: the custom build hook arranges for a short shell script to be called
instead of ghc. This shell script checks whether it is being called on HOC_cbits.o
and silently ignores that command. For all other parameters, it invokes the actual ghc.