On GNUstep, clean up the stack frame chain in Invocation.m:callWithExceptions
When the FFI call from Haskell to callWithExceptions() is made,
%ebp contains not a stack frame link but rather some other value.
When GNUstep's NSException tries to walk the stack to generate a
stack trace, it segfaults.
Therefore, we use GCC's __builtin_frame_address to retrieve
the saved frame pointer value, save it elsewhere and reset the
stack frame link to 0 so that GNUstep's stack trace just stops
there.
It is reset to the stored value upon exit.

Fixes issue #19.